FPLA-2951: fix merge conflict include copy of populator
