+++ service/src/test/java/uk/gov/hmcts/reform/fpl/service/orders/prepopulator/OrderSectionAndQuestionsPrePopulatorHolderTest.java
@@ -12,7 +12,6 @@ import uk.gov.hmcts.reform.fpl.service.orders.prepopulator.question.ApprovalDate
 import uk.gov.hmcts.reform.fpl.service.orders.prepopulator.question.ApprovalDateTimeBlockPrePopulator;
 import uk.gov.hmcts.reform.fpl.service.orders.prepopulator.question.ApproverBlockPrePopulator;
 import uk.gov.hmcts.reform.fpl.service.orders.prepopulator.question.EPOTypeAndPreventRemovalBlockPrePopulator;
-import uk.gov.hmcts.reform.fpl.service.orders.prepopulator.question.IssuingDetailsApprovalDateBlockPrePopulator;
 import uk.gov.hmcts.reform.fpl.service.orders.prepopulator.question.LinkedToHearingBlockPrePopulator;
 import uk.gov.hmcts.reform.fpl.service.orders.prepopulator.question.QuestionBlockOrderPrePopulator;
 import uk.gov.hmcts.reform.fpl.service.orders.prepopulator.question.WhichChildrenBlockPrePopulator;
@@ -57,7 +56,6 @@ class OrderSectionAndQuestionsPrePopulatorHolderTest {
     private ApprovalDateBlockPrePopulator approvalDateBlockPrePopulator;
     @Mock
     private ApprovalDateTimeBlockPrePopulator approvalDateTimeBlockPrePopulator;
-    private IssuingDetailsApprovalDateBlockPrePopulator issuingDetailsApprovalDateBlockPrePopulator;
 
     // Section blocks
     @Mock
@@ -88,8 +86,7 @@ class OrderSectionAndQuestionsPrePopulatorHolderTest {
             approvalDateTimeBlockPrePopulator,
             whichChildrenBlockPrePopulator,
             approverBlockPrePopulator,
-            epoTypeAndPreventRemovalBlockPrePopulator,
-            issuingDetailsApprovalDateBlockPrePopulator
+            epoTypeAndPreventRemovalBlockPrePopulator
         );
         questionBlockPrepopulatorMapping = Map.of(
             LINKED_TO_HEARING, linkedToHearingBlockPrePopulator,
@@ -97,8 +94,7 @@ class OrderSectionAndQuestionsPrePopulatorHolderTest {
             APPROVAL_DATE_TIME, approvalDateTimeBlockPrePopulator,
             APPROVER, approverBlockPrePopulator,
             WHICH_CHILDREN, whichChildrenBlockPrePopulator,
-            EPO_TYPE_AND_PREVENT_REMOVAL, epoTypeAndPreventRemovalBlockPrePopulator,
-            APPROVAL_DATE, issuingDetailsApprovalDateBlockPrePopulator
+            EPO_TYPE_AND_PREVENT_REMOVAL, epoTypeAndPreventRemovalBlockPrePopulator
         );
 
         sectionPrepopulatorMapping = Map.of(
